+++ test/unit/elastic_class/ElasticSearchRepo_create_repo.py
@@ -44,28 +44,8 @@ class Repo(object):
 
     Methods:
         get_repository -> Stub holder for snapshot.get_repository method.
-        create_repository -> Stub holder for snapshot.create_repository method.
 
     """
-
-    def create_repository(self, repository, body, verify):
-
-        """Method:  create_repository
-
-        Description:  Stub holder for snapshot.create_repository method.
-
-        Arguments:
-            (input) repository -> Name of repository to create.
-            (input) body -> Command for create.
-            (input) verify -> True|False - Validate creation.
-
-        """
-
-        if repository == "reponame3":
-            return {"acknowledged": False}
-
-        else:
-            return {"acknowledged": True}
 
     def get_repository(self):
 
@@ -89,7 +69,6 @@ class Elasticsearch(object):
 
     Methods:
         __init__ -> Initialize configuration environment.
-        ping -> Stub holder for Elasticsearch.ping method.
         info -> Stub holder for Elasticsearch.info method.
 
     """
@@ -106,21 +85,9 @@ class Elasticsearch(object):
 
         self.hosts = host_list
         self.port = port
-        self.ping_status = True
-        self.info_status = {"cluster_name": "ClusterName", "name": "servername"}
+        self.info_status = {"cluster_name": "ClusterName",
+                            "name": "servername"}
         self.snapshot = Repo()
-
-    def ping(self):
-
-        """Method:  ping
-
-        Description:  Stub holder for Elasticsearch.ping method.
-
-        Arguments:
-
-        """
-
-        return self.ping_status
 
     def info(self):
 
@@ -172,19 +139,25 @@ class UnitTest(unittest.TestCase):
             {"path": {"data": ["/dir/data1"], "logs": ["/dir/logs1"]}}},
             "serverid2": {"name": "hostname2", "settings":
             {"path": {"data": ["/dir/data2"], "logs": ["/dir/logs2"]}}}}
-        self.info_data = {"name": "localservername"}
         self.health_data = {"status": "green", "cluster_name": "ClusterName"}
-        self.master_name = "MasterName"
-        self.cluster_data = {"_nodes": {"total": 3}}
+        self.repo_list = {"reponame": {"type": "dbdump", "settings":
+                {"location": "/dir/path/dump"}}}
 
-    @mock.patch("elastic_class.get_cluster_nodes")
-    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.create_snapshot_repo",
+                mock.Mock(return_value={"acknowledged": False}))
+    @mock.patch("elastic_class.get_cluster_nodes",
+                mock.Mock(return_value={"_nodes": {"total": 3}}))
+    @mock.patch("elastic_class.get_master_name",
+                mock.Mock(return_value="MasterName"))
+    @mock.patch("elastic_class.get_info",
+                mock.Mock(return_value={"name": "localservername"}))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_not_created_repo(self, mock_es, mock_nodes, mock_info,
-                              mock_health, mock_master, mock_cluster):
+    def test_not_created_repo(self, mock_es, mock_nodes, mock_health,
+                              mock_list):
 
         """Function:  test_not_created_repo
 
@@ -196,10 +169,8 @@ class UnitTest(unittest.TestCase):
 
         mock_es.return_value = self.es
         mock_nodes.return_value = self.nodes_data
-        mock_info.return_value = self.info_data
         mock_health.return_value = self.health_data
-        mock_master.return_value = self.master_name
-        mock_cluster.return_value = self.cluster_data
+        mock_list.return_value = self.repo_list
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo,
                                              repo_dir=self.repo_dir)
@@ -208,14 +179,21 @@ class UnitTest(unittest.TestCase):
             (True,
             "ERROR:  Repository creation failure:  reponame3, /dir/path/repo"))
 
-    @mock.patch("elastic_class.get_cluster_nodes")
-    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.create_snapshot_repo",
+                mock.Mock(return_value={"acknowledged": True}))
+    @mock.patch("elastic_class.get_cluster_nodes",
+                mock.Mock(return_value={"_nodes": {"total": 3}}))
+    @mock.patch("elastic_class.get_master_name",
+                mock.Mock(return_value="MasterName"))
+    @mock.patch("elastic_class.get_info",
+                mock.Mock(return_value={"name": "localservername"}))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_not_detected_repo(self, mock_es, mock_nodes, mock_info,
-                               mock_health, mock_master, mock_cluster):
+    def test_not_detected_repo(self, mock_es, mock_nodes, mock_health,
+                               mock_list):
 
         """Function:  test_not_detected_repo
 
@@ -227,10 +205,8 @@ class UnitTest(unittest.TestCase):
 
         mock_es.return_value = self.es
         mock_nodes.return_value = self.nodes_data
-        mock_info.return_value = self.info_data
         mock_health.return_value = self.health_data
-        mock_master.return_value = self.master_name
-        mock_cluster.return_value = self.cluster_data
+        mock_list.return_value = self.repo_list
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo,
                                              repo_dir=self.repo_dir)
@@ -239,14 +215,21 @@ class UnitTest(unittest.TestCase):
             (True,
             "ERROR:  Repository not detected:  reponame2, /dir/path/repo"))
 
-    @mock.patch("elastic_class.get_cluster_nodes")
-    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.create_snapshot_repo",
+                mock.Mock(return_value={"acknowledged": False}))
+    @mock.patch("elastic_class.get_cluster_nodes",
+                mock.Mock(return_value={"_nodes": {"total": 3}}))
+    @mock.patch("elastic_class.get_master_name",
+                mock.Mock(return_value="MasterName"))
+    @mock.patch("elastic_class.get_info",
+                mock.Mock(return_value={"name": "localservername"}))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_missing_repo_name(self, mock_es, mock_nodes, mock_info,
-                               mock_health, mock_master, mock_cluster):
+    def test_missing_repo_name(self, mock_es, mock_nodes, mock_health,
+                               mock_list):
 
         """Function:  test_missing_repo_name
 
@@ -258,10 +241,8 @@ class UnitTest(unittest.TestCase):
 
         mock_es.return_value = self.es
         mock_nodes.return_value = self.nodes_data
-        mock_info.return_value = self.info_data
         mock_health.return_value = self.health_data
-        mock_master.return_value = self.master_name
-        mock_cluster.return_value = self.cluster_data
+        mock_list.return_value = self.repo_list
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo,
                                              repo_dir=self.repo_dir)
@@ -270,14 +251,20 @@ class UnitTest(unittest.TestCase):
             (True,
             "ERROR: Missing repo name or directory: 'None', '/dir/path/repo'"))
 
-    @mock.patch("elastic_class.get_cluster_nodes")
-    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.create_snapshot_repo",
+                mock.Mock(return_value={"acknowledged": True}))
+    @mock.patch("elastic_class.get_cluster_nodes",
+                mock.Mock(return_value={"_nodes": {"total": 3}}))
+    @mock.patch("elastic_class.get_master_name",
+                mock.Mock(return_value="MasterName"))
+    @mock.patch("elastic_class.get_info",
+                mock.Mock(return_value={"name": "localservername"}))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_no_repo_dir(self, mock_es, mock_nodes, mock_info, mock_health,
-                         mock_master, mock_cluster):
+    def test_no_repo_dir(self, mock_es, mock_nodes, mock_health, mock_list):
 
         """Function:  test_no_repo_dir
 
@@ -289,23 +276,27 @@ class UnitTest(unittest.TestCase):
 
         mock_es.return_value = self.es
         mock_nodes.return_value = self.nodes_data
-        mock_info.return_value = self.info_data
         mock_health.return_value = self.health_data
-        mock_master.return_value = self.master_name
-        mock_cluster.return_value = self.cluster_data
+        mock_list.return_value = self.repo_list
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo,
                                              repo_dir=self.repo_dir)
         self.assertEqual(es.create_repo(self.repo), (False, None))
 
-    @mock.patch("elastic_class.get_cluster_nodes")
-    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.create_snapshot_repo",
+                mock.Mock(return_value={"acknowledged": True}))
+    @mock.patch("elastic_class.get_cluster_nodes",
+                mock.Mock(return_value={"_nodes": {"total": 3}}))
+    @mock.patch("elastic_class.get_master_name",
+                mock.Mock(return_value="MasterName"))
+    @mock.patch("elastic_class.get_info",
+                mock.Mock(return_value={"name": "localservername"}))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_no_repo_name(self, mock_es, mock_nodes, mock_info, mock_health,
-                          mock_master, mock_cluster):
+    def test_no_repo_name(self, mock_es, mock_nodes, mock_health, mock_list):
 
         """Function:  test_no_repo_name
 
@@ -317,24 +308,28 @@ class UnitTest(unittest.TestCase):
 
         mock_es.return_value = self.es
         mock_nodes.return_value = self.nodes_data
-        mock_info.return_value = self.info_data
         mock_health.return_value = self.health_data
-        mock_master.return_value = self.master_name
-        mock_cluster.return_value = self.cluster_data
+        mock_list.return_value = self.repo_list
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo,
                                              repo_dir=self.repo_dir)
         self.assertEqual(es.create_repo(repo_dir=self.repo_dir),
                          (False, None))
 
-    @mock.patch("elastic_class.get_cluster_nodes")
-    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.create_snapshot_repo",
+                mock.Mock(return_value={"acknowledged": True}))
+    @mock.patch("elastic_class.get_cluster_nodes",
+                mock.Mock(return_value={"_nodes": {"total": 3}}))
+    @mock.patch("elastic_class.get_master_name",
+                mock.Mock(return_value="MasterName"))
+    @mock.patch("elastic_class.get_info",
+                mock.Mock(return_value={"name": "localservername"}))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_default(self, mock_es, mock_nodes, mock_info, mock_health,
-                     mock_master, mock_cluster):
+    def test_default(self, mock_es, mock_nodes, mock_health, mock_list):
 
         """Function:  test_default
 
@@ -346,10 +341,8 @@ class UnitTest(unittest.TestCase):
 
         mock_es.return_value = self.es
         mock_nodes.return_value = self.nodes_data
-        mock_info.return_value = self.info_data
         mock_health.return_value = self.health_data
-        mock_master.return_value = self.master_name
-        mock_cluster.return_value = self.cluster_data
+        mock_list.return_value = self.repo_list
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo,
                                              repo_dir=self.repo_dir)