+++ CHANGELOG.md
@@ -8,9 +8,14 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 Breaking Change
 
 ### Changed
+- elastic_class.ElasticSearchRepo.delete_repo:  Replaced snapshot.delete_repository with call to delete_snapshot_repo().
+- elastic_class.ElasticSearchRepo.create_repo:  Replaced es.snapshot.get_repository with call to get_repo_list().
+- elastic_class.ElasticSearchDump.dump_db:  Replaced snapshot.create with call to create_snapshot().
+- elastic_class.ElasticSearchRepo.create_repo:  Replaced snapshot.create_repository with call to create_snapshot_repo().
+- elastic_class.ElasticSearch.update_status:  Replaced call to ping with call to is_active().
 - elastic_class.ElasticSearch.\_\_init\_\_:  Integrated ElasticCluster class into ElasticSearch class.
 - elastic_class.ElasticSearch.\_\_init\_\_:  Integrated Elastic class into ElasticSearch class.
-- elastic_class.ElasticSearch.\_\_init\_\_:  Replaced attribute updates with call to update_status.
+- elastic_class.ElasticSearch.\_\_init\_\_:  Replaced attribute updates with call to update_status().
 
 ### Added
 - elastic_class.ElasticSearchStatus:  Added class and associated methods.